Move Synapse tests to separated directories
