+++ tests/Functional/Table/Synapse/SynapseTableReflectionTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Keboola\TableBackendUtils\Functional\Table;
+namespace Tests\Keboola\TableBackendUtils\Functional\Table\Synapse;
 
 use Generator;
 use Keboola\TableBackendUtils\Column\ColumnCollection;
